move tests to k8s file
